Began adding flash messaging, fixed recipe edit, additional user config

+++ src/main/java/com/recipes/web/controller/RecipeController.java
@@ -7,17 +7,21 @@ import com.recipes.dao.UserRepository;
 import com.recipes.model.Category;
 import com.recipes.model.Recipe;
 import com.recipes.model.User;
+import com.recipes.web.FlashMessage;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static com.recipes.web.FlashMessage.Status.FAILURE;
+import static com.recipes.web.FlashMessage.Status.SUCCESS;
 
 @Controller
 public class RecipeController {
@@ -37,6 +41,8 @@ public class RecipeController {
 
     @RequestMapping("/recipes")
     public String index(Model model) {
+        User user = (User) model.asMap().get("currentUser");
+        model.addAttribute("authenticated", user.isAdmin());
         List<Category> allCategories = (List<Category>) categoryRepository.findAll();
         model.addAttribute("categories", allCategories);
         List<Recipe> allRecipes = (List<Recipe>) recipeRepository.findAll();
@@ -52,15 +58,22 @@ public class RecipeController {
     }
 
     @RequestMapping(path = "/recipes/{id}/delete", method = RequestMethod.POST)
-    public String deleteRecipe(@PathVariable("id") int id) {
+    public String deleteRecipe(@PathVariable("id") int id, Model model, RedirectAttributes attributes) {
+        User user = (User) model.asMap().get("currentUser");
         Recipe recipe = recipeRepository.findOne((long) id);
-        recipe.getCreatedBy().removeCreatedRecipe(recipe);
-        users.save(recipe.getCreatedBy());
-        recipe.setCreatedBy(null);
-        recipeRepository.delete((long) id);
+        if(recipe.getCreatedBy() == user || user.isAdmin()) {
+            recipe.getCreatedBy().removeCreatedRecipe(recipe);
+            users.save(recipe.getCreatedBy());
+            recipe.setCreatedBy(null);
+            recipeRepository.delete((long) id);
+            attributes.addFlashAttribute("flash", new FlashMessage("Recipe deleted successfully", SUCCESS));
+        } else {
+            attributes.addFlashAttribute("flash", new FlashMessage("You are not allowed to delete this recipe.", FAILURE));
+        }
         return "redirect:/";
     }
 
+    //TODO: user-admin
     @RequestMapping("/recipes/{id}/edit")
     public String edit(Model model, @PathVariable("id") int id) {
         Recipe recipe = recipeRepository.findOne((long) id);
@@ -76,7 +89,7 @@ public class RecipeController {
         List<Recipe> queriedRecipes = new ArrayList<>();
         if(searchQuery != null) {
             model.addAttribute("search", searchQuery);
-            List<Recipe> namedRecipes = recipeRepository.findByNameContaining(searchQuery);
+            List<Recipe> namedRecipes = recipeRepository.findByDescriptionContaining(searchQuery);
             if(!category.equals("")) {
                 List<Recipe> categorizedRecipes = recipeRepository.findByCategoryName(category);
                 queriedRecipes = namedRecipes.stream().filter(categorizedRecipes::contains).collect(Collectors.toList());
@@ -110,7 +123,7 @@ public class RecipeController {
         }
         recipe.getIngredients().forEach(ingredient -> ingredientRepository.save(ingredient));
         if(recipe.getCreatedBy() == null) {
-            User user = (User)model.asMap().get("user");
+            User user = (User)model.asMap().get("currentUser");
             recipe.setCreatedBy(user);
             recipeRepository.save(recipe);
             users.save(user);
@@ -129,7 +142,13 @@ public class RecipeController {
             categoryRepository.save(category);
         }
         recipe.getIngredients().forEach(ingredient -> ingredientRepository.save(ingredient));
+        recipe.getInstructions()
+                .stream()
+                .filter(instruction -> instruction.equals(""))
+                .collect(Collectors.toList())
+                .forEach(recipe::removeInstruction);
         recipeRepository.save(recipe);
         return "redirect:/recipes/" + recipe.getId();
     }
+
 }
